+++ changelog.md
@@ -1,4 +1,17 @@
 # 0.1.10-dev
+
+## Specification 38 - 2021-09-20
+### Added
+- Created and added pallet `ternoa-sgx`
+### Changed
+- types.json
+- ternoa-nfts weights
+### Removed
+- Removed pallet `pallet-substratee-registry`
+
+## Specification 37 - 2021-09-11
+Skipped
+
 ## Specification 36 - 2021-09-10
 ### Added
 - Marketplaces now have names (Storage migration needed)